[af-326] chore: Stock Controller, Service 생성 후 분리
