+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/restaurant/restaurant/application/AdminRestaurantService.java
@@ -1,9 +1,8 @@
 package com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.application;
 
+import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.request.RestaurantRequest;
 import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.response.RestaurantDetailResponse;
 import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.response.RestaurantListResponse;
-import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.request.RestaurantRequest;
-import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.response.RestaurantStockListResponse;
 import com.drunkenlion.alcoholfriday.domain.member.entity.Member;
 import org.springframework.data.domain.Page;
 
@@ -17,6 +16,4 @@ public interface AdminRestaurantService {
     RestaurantDetailResponse modifyRestaurant(Member member, Long id, RestaurantRequest restaurantRequest);
 
     void deleteRestaurant(Member member, Long id);
-
-    Page<RestaurantStockListResponse> getRestaurantStocks(Member member, Long restaurantId, int page, int size);
 }
