+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/restaurant/restaurant/application/AdminRestaurantServiceImpl.java
@@ -3,10 +3,8 @@ package com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.applica
 import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.request.RestaurantRequest;
 import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.response.RestaurantDetailResponse;
 import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.response.RestaurantListResponse;
-import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.response.RestaurantStockListResponse;
 import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.dto.response.RestaurantStockProductResponse;
 import com.drunkenlion.alcoholfriday.domain.admin.restaurant.restaurant.util.RestaurantDataValidator;
-import com.drunkenlion.alcoholfriday.domain.admin.restaurant.util.RestaurantValidator;
 import com.drunkenlion.alcoholfriday.domain.member.dao.MemberRepository;
 import com.drunkenlion.alcoholfriday.domain.member.entity.Member;
 import com.drunkenlion.alcoholfriday.domain.member.enumerated.MemberRole;
@@ -164,21 +162,6 @@ public class AdminRestaurantServiceImpl implements AdminRestaurantService {
                 .build();
 
         restaurantRepository.save(restaurant);
-    }
-
-    @Override
-    public Page<RestaurantStockListResponse> getRestaurantStocks(Member member, Long restaurantId, int page, int size) {
-        Restaurant restaurant = restaurantRepository.findByIdAndDeletedAtIsNull(restaurantId)
-                .orElseThrow(() -> new BusinessException(HttpResponse.Fail.NOT_FOUND_RESTAURANT));
-
-        if (member.getRole().equals(MemberRole.OWNER)) {
-            RestaurantValidator.validateOwnership(member, restaurant);
-        }
-
-        Pageable pageable = PageRequest.of(page, size);
-        Page<RestaurantStock> stockPages = restaurantStockRepository.findRestaurantStocks(member, restaurant, pageable);
-
-        return stockPages.map(stock -> RestaurantStockListResponse.of(stock, fileService.findOne(stock.getProduct())));
     }
 
     private List<RestaurantStockProductResponse> getRestaurantStockItemResponseList(Restaurant restaurant) {
